+++ src/Poc.Mobile.App/IBaseUrl.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace Poc.Mobile.App
+{
+    public interface IBaseUrl
+    {
+        string Get();
+    }
+}