+++ src/Poc.Mobile.App/Views/Legal/LegalPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+
+using Xamarin.Forms;
+
+namespace Poc.Mobile.App.Views.Legal
+{
+    public partial class LegalPage : ContentPage
+    {
+        public LegalPage()
+        {
+            InitializeComponent();
+        }
+    }
+}